packages/b2b_ec_sources/pyproject.toml
[project]
name = "b2b-ec-sources"
version = "0.1.0"
description = "High-performance data simulation suite for B2B E-commerce source environments, including relational DB generation and weblog synthesis."
dependencies = [
    "psycopg2-binary",
    "faker",
    "polars",
    "numpy",
    "pycountry",
    "b2b-ec-utils",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/b2b_ec_sources"]

[tool.uv.sources]
b2b-ec-utils = { workspace = true, editable = true }
